Add status monitoring for new raw yaml step
